Rename photon to point cloud
ATL07 points are not actually photons, but an aggregate of ATL03 points.
