+++ book/_toc.yml
@@ -38,8 +38,8 @@ parts:
             - file: tutorials/cloud-computing/atl08_parquet_files/atl08_parquet
               options:
                 - titlesonly: true
-        - file: tutorials/mental-health/index          
-        - file: tutorials/machine-learning/photon_classifier.ipynb
+        - file: tutorials/mental-health/index
+        - file: tutorials/machine-learning/point_cloud_classifier.ipynb
 - caption: Projects
   chapters:
     - file: projects/index
