+++ book/tutorials/index.md
@@ -10,4 +10,4 @@ Below you'll find a table keeping track of all tutorials presented at this event
 | [ICESat-2 Mission](./mission-overview/icesat-2-mission-overview.ipynb) | ICESat-2 Mission and Products | n/a |  Not recorded |
 | [Cloud Computing](./cloud-computing/00-goals-and-outline.ipynb) | Cloud Computing Tutorial | n/a |  Not recorded |
 | [Notebooks to Packages](./nb-to-package/index.md) | All about Python classes to packages | n/a |  Not recorded |
-| [ICESat-2 photon classification](./machine-learning/photon_classifier.ipynb) | Machine Learning, PyTorch | ATL07 | Not recorded |
+| [ICESat-2 point cloud classification](./machine-learning/point_cloud_classifier.ipynb) | Machine Learning, PyTorch | ATL07 | Not recorded |
